feat(lightning): Enable LNURL
Enables lnurl logic in the scanner switch statement.
Moves lnurl scanner logic to lnurl.ts.
Adds getPeersFromStorage method to utils/lightning/index.ts.
